break out get-layers as separate function
